backend/requirements.txt
Django==4.2
djangorestframework
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt
Pillow
python-dotenv
scikit-learn
numpy
joblib
opencv-python
ultralytics
mediapipe
scipy
